Move http_util to raw
