+++ src/services/azblob/backend.rs
@@ -30,7 +30,6 @@ use reqsign::AzureStorageSigner;
 
 use super::dir_stream::DirStream;
 use super::error::parse_error;
-use crate::http_util::*;
 use crate::object::ObjectMetadata;
 use crate::object::ObjectPager;
 use crate::raw::*;
