+++ src/services/gcs/backend.rs
@@ -34,7 +34,6 @@ use super::dir_stream::DirStream;
 use super::error::parse_error;
 use super::error::parse_json_deserialize_error;
 use super::uri::percent_encode_path;
-use crate::http_util::*;
 use crate::object::ObjectPager;
 use crate::raw::*;
 use crate::*;
