+++ src/services/gcs/dir_stream.rs
@@ -23,15 +23,9 @@ use time::OffsetDateTime;
 use super::backend::Backend;
 use super::error::parse_error;
 use super::error::parse_json_deserialize_error;
-use crate::http_util::parse_error_response;
 use crate::object::ObjectPage;
 use crate::raw::*;
-use crate::Error;
-use crate::ErrorKind;
-use crate::ObjectEntry;
-use crate::ObjectMetadata;
-use crate::ObjectMode;
-use crate::Result;
+use crate::*;
 
 /// DirStream takes over task of listing objects and
 /// helps walking directory
