+++ src/services/ipmfs/dir_stream.rs
@@ -21,7 +21,6 @@ use serde::Deserialize;
 use super::backend::Backend;
 use super::error::parse_error;
 use super::error::parse_json_deserialize_error;
-use crate::http_util::parse_error_response;
 use crate::object::ObjectPage;
 use crate::raw::*;
 use crate::ObjectEntry;
