+++ src/services/obs/backend.rs
@@ -27,7 +27,6 @@ use log::debug;
 use reqsign::HuaweicloudObsSigner;
 
 use super::error::parse_error;
-use crate::http_util::*;
 use crate::object::ObjectPager;
 use crate::raw::*;
 use crate::services::obs::dir_stream::DirStream;
