+++ src/services/s3/backend.rs
@@ -40,7 +40,6 @@ use serde::Serialize;
 use super::dir_stream::DirStream;
 use super::error::parse_error;
 use super::error::parse_xml_deserialize_error;
-use crate::http_util::*;
 use crate::object::ObjectPager;
 use crate::raw::*;
 use crate::*;
